+++ game/gamesys/Class.cpp
@@ -447,9 +447,9 @@ void * idClass::operator new( size_t s ) {
 	int *p;
 
 	s += sizeof( int );
-	p = (int *)Mem_Alloc( s );
-	*p = s;
-	memused += s;
+    p = (int *)Mem_Alloc(static_cast<int>(s));
+    *p = static_cast<int>(s);
+    memused += static_cast<int>(s);
 	numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
@@ -469,9 +469,9 @@ void * idClass::operator new( size_t s, int, int, char *, int ) {
 	int *p;
 
 	s += sizeof( int );
-	p = (int *)Mem_Alloc( s );
-	*p = s;
-	memused += s;
+    p = (int *)Mem_Alloc(static_cast<int>(s));
+    *p = static_cast<int>(s);
+    memused += static_cast<int>(s);
 	numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
